refactoring code, removing JwtConfig class whose beans are not really needed here
