--- src/main/java/fr/patedor/PFR_Equipe/security/jwt/JwtConfig.java
@@ -1,42 +1,0 @@
-package fr.patedor.PFR_Equipe.security.jwt;
-
-import fr.patedor.PFR_Equipe.repository.EmployeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-
-@Configuration
-public class JwtConfig {
-
-    @Autowired
-    private EmployeRepository employeRepository;
-
-    //Instancie un userDetailsService, qui est juste un DAO qui a accès uniquement à findByLogin
-    @Bean
-    UserDetailsService userDetailsService() {
-        return login -> employeRepository.findByLogin(login)
-                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
-    }
-
-    @Bean
-    AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService());
-        authProvider.setPasswordEncoder(passwordEncoder());
-        return authProvider;
-    }
-
-    //Encodeur de password avec BCrypt, car c'est le standard avec Spring Security
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
